DVFS/PSU: use uint32_t instead of uint64_t for voltage variable.
In this patch voltage variable is change from 64 to 32 bits.
This resolution is more than enough for this implementation.

+++ module/dvfs/include/mod_dvfs.h
@@ -38,7 +38,7 @@ struct mod_dvfs_frequency_limits {
  * \brief Operating Performance Point (OPP).
  */
 struct mod_dvfs_opp {
-    uint64_t voltage; /*!< Power supply voltage in millivolts (mV) */
+    uint32_t voltage; /*!< Power supply voltage in millivolts (mV) */
     uint32_t frequency; /*!< Clock rate in Hertz (Hz) */
     uint32_t power; /*!< Power draw in milliwatts(mW) */
 };
